requirements.txt
alembic==1.16.4
APIFlask==2.4.0
apispec==6.8.2
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Deprecated==1.2.18
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
psycopg2==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.1.1
rich==13.9.4
SQLAlchemy==2.0.42
typing_extensions==4.14.1
webargs==8.7.0
Werkzeug==3.1.3
wrapt==1.17.2

# Testing dependencies
pytest==8.3.4
pytest-cov==6.0.0
pytest-flask==1.3.0
coverage==7.6.9
